wxGUI: new InfoManagerDataCatalog class created and _onLearnMore was placed in this class.
