+++ gui/wxpython/datacatalog/__init__.py
@@ -2,5 +2,6 @@ all = [
     'catalog',
     'frame',
     'tree',
-    'dialogs'
+    'dialogs',
+    'infomanager'
 ]
