+++ gui/wxpython/datacatalog/catalog.py
@@ -17,12 +17,11 @@ for details.
 
 import wx
 import os
-import webbrowser
 
 from core.debug import Debug
 from datacatalog.tree import DataCatalogTree
 from datacatalog.toolbars import DataCatalogToolbar
-from gui_core.infobar import InfoManager
+from datacatalog.infomanager import InfoManagerDataCatalog
 
 from grass.pydispatch.signal import Signal
 
@@ -62,11 +61,11 @@ class DataCatalog(wx.Panel):
                   flag=wx.EXPAND)
 
         #infobar instance
-        self.infoManager = InfoManager(guiparent=self, sizer=self.sizer)
+        self.infoManager = InfoManagerDataCatalog(guiparent=self, sizer=self.sizer)
 
         # Show infobar for first-time user
         if is_current_mapset_in_demolocation:
-            buttons1 = [("Learn More", self._onLearnMore)]
+            buttons1 = [("Learn More", self.infoManager._onLearnMore)]
             buttons2 = [("Create new Location", self.OnCreateLocation)]
             self.infoManager.ShowInfoBar1(buttons1)
             self.infoManager.ShowInfoBar2(buttons2)
@@ -75,10 +74,6 @@ class DataCatalog(wx.Panel):
         self.SetSizer(self.sizer)
 
         self.Layout()
-
-    def _onLearnMore(self, event):
-        webbrowser.open("https://grass.osgeo.org/grass79/manuals/grass_database.html")
-        event.Skip()
 
     def LoadItems(self):
         self.tree.ReloadTreeItems()
@@ -109,6 +104,7 @@ class DataCatalog(wx.Panel):
         """Create new location"""
         db_node, loc_node, mapset_node = self.tree.GetCurrentDbLocationMapsetNode()
         self.tree.CreateLocation(db_node)
+        event.Skip()
 
     def OnDownloadLocation(self, event):
         """Download location to current grass database"""
